[Documents] Open vs. Preview Button - resolves #3185
