+++ bundles/AdminBundle/Resources/public/js/pimcore/document/tree.js
@@ -630,8 +630,8 @@ pimcore.document.tree = Class.create({
 
         if ((record.data.type == "page" || record.data.type == "hardlink") && record.data.permissions.view && perspectiveCfg.inTreeContextMenu("document.open")) {
             menu.add(new Ext.menu.Item({
-                text: t('open'),
-                iconCls: "pimcore_icon_cursor",
+                text: t('open_in_new_window'),
+                iconCls: "pimcore_icon_open_window",
                 handler: function () {
                     window.open(record.data.url);
                 }.bind(this)
